feat: added haarcascade detector
