+++ .gitignore
@@ -127,3 +127,9 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+#.idea path
+.idea/
+
+#lock files
+*.lock
