+++ face-analysis/app/infrastructure/repositories/detector/models.py
@@ -1,0 +1,8 @@
+from cv2 import CascadeClassifier
+
+
+class CascadeDetectorModel:
+
+    @staticmethod
+    def load(path: str) -> CascadeClassifier:
+        return CascadeClassifier(path)
