+++ face-analysis/docs/roadmap.md
@@ -1,0 +1,11 @@
+# Face Analysis Roadmap 
+
+- [X] Disponibilizar modelos pré-treinados;
+- [X] Criar projeto base;
+- [X] Criar gerenciador de configurações;
+- [X] Criar repositorio referente a captura de imagens da câmera;
+- [X] Criar detector com o haarcascade;
+- [ ] Criar detector com o deep learning;
+- [X] Criar detector em real time com o haarcascade;
+- [ ] Criar detector em real time com o deep learning;
+- [ ] Utilizar o mesmo repository para diferentes abordagens de entrada (tradicional ou deep learning);
